requirements.txt
decorator==4.0.4
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
requests==2.8.1
requests-oauthlib==0.5.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
tweepy==3.4.0
Unidecode==0.4.18
wheel==0.24.0
pymongo==3.0.3
celery-with-redis==3.0
nltk==3.1
unicodecsv==0.14.1
textblob==0.11.0
python-dateutil==2.5.3